refactor(ClassExpression): changed function getClassExpressionJSX to component ClassExpression
